refactor: 모달 파일 경로 수정
