+++ src/components/main/restaurant/RestaurantList.jsx
@@ -5,18 +5,17 @@ const RestaurantList = ({ restaurants, onCardClick }) => {
   return (
     <section className={`${styles.restaurantListContainer}`}>
       <ul className={`${styles.restaurantList}`}>
-        {restaurants.length > 0 &&
-          restaurants.map((item) => {
-            return (
-              <RestaurantCard
-                key={item.id}
-                alt={item.category}
-                name={item.name}
-                description={item.description}
-                onCardClick={onCardClick}
-              />
-            );
-          })}
+        {restaurants.map((item) => {
+          return (
+            <RestaurantCard
+              key={item.id}
+              alt={item.category}
+              name={item.name}
+              description={item.description}
+              onCardClick={onCardClick}
+            />
+          );
+        })}
       </ul>
     </section>
   );
